requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
jinja2==3.1.5
python-dotenv==1.0.1
aiosqlite>=0.20
apscheduler>=3.10
